Move cell manager strategies
